ove sequential notifier to own file
