+++ sqlitecluster/SQLiteNode.h
@@ -1,29 +1,9 @@
 #pragma once
 #include "SQLite.h"
+#include "SQLiteSequentialNotifier.h"
 #include "WallClockTimer.h"
 class SQLiteCommand;
 class SQLiteServer;
-
-class SequentialNotifier {
-  public:
-    SequentialNotifier() : _value(0) {}
-    bool waitFor(uint64_t value);
-    void notifyThrough(uint64_t value);
-    void cancel();
-
-  private:
-    struct WaitState {
-        WaitState() : canceled(false), completed(false) {}
-        mutex m;
-        condition_variable cv;
-        bool canceled;
-        bool completed;
-    };
-
-    mutex _m;
-    map<uint64_t, shared_ptr<WaitState>> _pending;
-    uint64_t _value;
-};
 
 // Distributed, leader/follower, failover, transactional DB cluster
 class SQLiteNode : public STCPNode {
@@ -256,8 +236,8 @@ class SQLiteNode : public STCPNode {
     set<string> _replicationHashesToRollback;
 
     mutex _replicationMutex;
-    SequentialNotifier _dbNotifier;
-    SequentialNotifier _commitNotifier;
+    SQLiteSequentialNotifier _dbNotifier;
+    SQLiteSequentialNotifier _commitNotifier;
 
     // Replication thread main body.
     static void replicate(SQLiteNode& node, Peer* peer, SData command);
